Fix bug of Vehicle destructor hanging
To reproduce the bug on Linux, make sure the host is NOT connected to the A3 flight controller, then run the Linux telemetry sample.

The output will say "Comms appear to be incorrectly set up. Exiting." but it will hang instead of exiting.

The location of the hang was in Vehicle's destructor, where it calls "this->readThread->stopThread();"

This commit fixes the issue.